Fix warning, cast return value from pthread_self to unsigned long On Mac OS X pthread_self returns a pointer, openssl has support for using pointers but their own example code casts the value too.

+++ src/networking/net_posix.c
@@ -50,7 +50,7 @@ static pthread_mutex_t *ssl_locks;
 static unsigned long
 ssl_tid_fn(void)
 {
-  return pthread_self();
+  return (unsigned long)pthread_self();
 }
 
 static void
